normalise / refactore functions

+++ ft_printf2/src/ft_char.c
@@ -6,13 +6,13 @@
 /*   By: skydogzz </var/spool/mail/skydogzz>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/17 15:31:39 by skydogzz          #+#    #+#             */
-/*   Updated: 2024/11/17 15:31:59 by skydogzz         ###   ########.fr       */
+/*   Updated: 2024/11/17 16:24:41 by skydogzz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/ft_printf.h"
 
-int	ft_putcharoptions_fd(const char c, int fd)
+size_t	ft_putcharoptions_fd(const char c, int fd)
 {
 	ft_putchar_fd(c, 1);
 	return (1);
